+++ quantforge/__init__.py
@@ -779,6 +779,7 @@ from .cur import cur, column_leverage_scores
 from .cp_decomp import cp_als, cp_reconstruct
 from .tucker import tucker_hosvd, tucker_reconstruct
 from .spectral_clustering import spectral_clustering, rbf_affinity
+from .mds import classical_mds, embedded_distances
 from .dual import Dual, derivative as dual_derivative
 from .dual_calculus import dual_gradient, dual_newton
 from .hyperdual import HyperDual, derivatives as hyperdual_derivatives, second_derivative
@@ -1483,7 +1484,7 @@ try:  # pragma: no cover - trivial availability branch
 except ImportError:  # pragma: no cover
     HAS_NUMPY = False
 
-__version__ = "5.72.0"
+__version__ = "5.73.0"
 
 __all__ = [
     "OptionType",
@@ -2535,6 +2536,8 @@ __all__ = [
     "tucker_reconstruct",
     "spectral_clustering",
     "rbf_affinity",
+    "classical_mds",
+    "embedded_distances",
     "Dual",
     "dual_derivative",
     "dual_gradient",